refactor: split start/stop functions
