+++ packages/backend/src/handlers/api/stop_pulse/requirements.txt
@@ -1,0 +1,4 @@
+aws-lambda-powertools[parser]
+aws_xray_sdk
+boto3
+pydantic
