+++ packages/infrastructure/lib/lambda-stack.ts
@@ -9,14 +9,15 @@ import * as path from 'path';
 
 interface LambdaStackProps extends cdk.StackProps {
     startPulseTable: dynamodb.ITable;
-    stopPulseTableV1: dynamodb.ITable;
+    stopPulseTable: dynamodb.ITable;
     ingestedPulseTable: dynamodb.ITable;
     pulsesIngestionQueue: sqs.IQueue;
     pulsesIngestionQueueDLQ: sqs.IQueue;
 }
 
 export class LambdaStack extends cdk.Stack {
-    public readonly pythonStartStopFunction: PythonFunction;
+    public readonly pythonStartFunction: PythonFunction;
+    public readonly pythonStopFunction: PythonFunction;
     public readonly pythonIngestPulseFunction: PythonFunction;
 
     constructor(scope: Construct, id: string, props: LambdaStackProps) {
@@ -28,63 +29,71 @@ export class LambdaStack extends cdk.Stack {
             description: 'Shared dependencies layer',
         });
 
-        this.pythonStartStopFunction = new PythonFunction(this, 'StartPulse', {
-            entry: path.resolve('../backend/src/handlers/api/start_pulse'),
+        const bundlingAssetExcludes = [
+            '**/__pycache__',
+            '**/*.pyc',
+            '**/*.pyo',
+            '**/*.pyd',
+        ]
+
+        const commonLambdaProps = {
             index: 'app.py',
             handler: 'handler',
             architecture: lambda.Architecture.ARM_64,
             runtime: lambda.Runtime.PYTHON_3_13,
-            timeout: cdk.Duration.seconds(5),
-            functionName: 'ps-start-stop-pulse',
-            description: 'Function to start & stop pulses',
-            logRetention: cdk.aws_logs.RetentionDays.FIVE_DAYS,
-
             layers: [sharedLayer],
+            bundling: {
+                assetExcludes: bundlingAssetExcludes
+            },
+            timeout: cdk.Duration.seconds(5),
+        }
+
+        this.pythonStartFunction = new PythonFunction(this, 'StartPulseV1', {
+            ...commonLambdaProps,
+            entry: path.resolve('../backend/src/handlers/api/start_pulse'),
+            functionName: 'ps-start-pulse',
+            description: 'Function to start pulses',
+            logRetention: cdk.aws_logs.RetentionDays.FIVE_DAYS,
             environment: {
                 START_PULSE_TABLE_NAME: props.startPulseTable.tableName,
-                STOP_PULSE_TABLE_NAME: props.stopPulseTableV1.tableName,
-            },
-            bundling: {
-                assetExcludes: [
-                    '**/__pycache__',
-                    '**/*.pyc',
-                    '**/*.pyo',
-                    '**/*.pyd',
-                ]
             },
         });
 
-        props.startPulseTable.grantReadWriteData(this.pythonStartStopFunction);
-        props.stopPulseTableV1.grantReadWriteData(this.pythonStartStopFunction);
+        this.pythonStopFunction = new PythonFunction(this, 'StopPulse', {
+            entry: path.resolve('../backend/src/handlers/api/stop_pulse'),
+            ...commonLambdaProps,
+            functionName: 'ps-stop-pulse',
+            description: 'Function to stop pulses',
+            environment: {
+                START_PULSE_TABLE_NAME: props.startPulseTable.tableName,
+                STOP_PULSE_TABLE_NAME: props.stopPulseTable.tableName,
+            },
+        });
+
+        props.startPulseTable.grantReadWriteData(this.pythonStartFunction);
+        props.startPulseTable.grantReadWriteData(this.pythonStopFunction);
+        props.stopPulseTable.grantReadWriteData(this.pythonStopFunction);
 
         this.pythonIngestPulseFunction = new PythonFunction(this, 'IngestPulse', {
+            ...commonLambdaProps,
             entry: path.resolve('../backend/src/handlers/events/ingest_pulse'),
-            index: 'app.py',
-            handler: 'handler',
-            architecture: lambda.Architecture.ARM_64,
-            runtime: lambda.Runtime.PYTHON_3_13,
             timeout: cdk.Duration.seconds(60),
             functionName: 'ps-ingest-pulse',
             description: 'Function to ingest pulses',
             logRetention: cdk.aws_logs.RetentionDays.FIVE_DAYS,
             layers: [sharedLayer],
             environment: {
-                STOP_PULSE_TABLE_NAME: props.stopPulseTableV1.tableName,
+                STOP_PULSE_TABLE_NAME: props.stopPulseTable.tableName,
                 INGESTED_PULSE_TABLE_NAME: props.ingestedPulseTable.tableName,
                 SQS_QUEUE_ARN: props.pulsesIngestionQueue.queueArn,
                 SQS_DLQ_ARN: props.pulsesIngestionQueueDLQ.queueArn,
             },
             bundling: {
-                assetExcludes: [
-                    '**/__pycache__/**',
-                    '**/*.pyc',
-                    '**/*.pyo',
-                    '**/*.pyd',
-                ]
+                assetExcludes: bundlingAssetExcludes
             },
         });
 
-        props.stopPulseTableV1.grantWriteData(this.pythonIngestPulseFunction);
+        props.stopPulseTable.grantWriteData(this.pythonIngestPulseFunction);
         props.ingestedPulseTable.grantWriteData(this.pythonIngestPulseFunction);
         props.pulsesIngestionQueue.grantConsumeMessages(this.pythonIngestPulseFunction);
